Squashed commit of the following:
commit 5dc7acc6582e158bacb9ce4618ff97a23c869ddc
Merge: 7e339db b75a178
Author: tmdwnsasa <tmdwns_as@naver.com>
Date:   Wed May 1 20:44:54 2024 +0900

    Merge pull request #6 from JHLee0891/movieReview

    5월 1일 마지막 커밋

commit b75a1780d8d969ad928629fa0dea892019cfd3ad
Author: tmdwnsasa <tmdwns_as@naver.com>
Date:   Wed May 1 20:40:29 2024 +0900

    5월 1일 마지막 커밋
    팀 프로젝트 작업물 merge

commit 7e339dbfedc1129b64c0d8f4cdce5e8d5433c8d6
Author: soooomm00@gmail.com <soooomm00@gmail.com>
Date:   Wed May 1 20:15:30 2024 +0900

    fix : 충돌 해결

commit 6850de7e98ee19c802dfac25d915cf0257ac0756
Merge: 49d3884 1bed318
Author: SooooMm <67359937+SooooMm@users.noreply.github.com>
Date:   Wed May 1 20:06:05 2024 +0900

    Merge pull request #5 from JHLee0891/Feature/#4-detailPage-UI

    [Feat] 상세페이지 구현

commit 1bed318d764f9d749c04e1387fbaaafd0a33fae2
Merge: 8f33120 49d3884
Author: soooomm00@gmail.com <soooomm00@gmail.com>
Date:   Wed May 1 20:04:34 2024 +0900

    Merge branch 'main' into Feature/#4-detailPage-UI

commit 8f331201a99bde6915349a8ef26a58803315f00a
Author: soooomm00@gmail.com <soooomm00@gmail.com>
Date:   Wed May 1 19:46:27 2024 +0900

    feat : 상세페이지 UI 구현 및 폴더 구조 변경

commit 49d3884bc453be9459f057bc5310107422dbbaf5
Merge: adf41aa 5eb91db
Author: NohSiHeon <91312161+NohSiHeon@users.noreply.github.com>
Date:   Wed May 1 19:21:36 2024 +0900

    Merge pull request #3 from JHLee0891/Feature/#1

    상세페이지 API 추가

commit 5eb91db9df00d7612041c4bd7d3799839b5f6dd4
Merge: c909542 adf41aa
Author: NohSiHeon <shtlgjs016@naver.com>
Date:   Wed May 1 19:17:47 2024 +0900

    상세페이지 API 추가

commit c909542f32e86d1d0a67fc622294c4d998f6cdbc
Author: NohSiHeon <shtlgjs016@naver.com>
Date:   Wed May 1 19:03:26 2024 +0900

    상세페이지 API 추가

commit adf41aaba38d74c03ddb94d737f4472baa4eb12a
Merge: e0da24a 593abe4
Author: Poison6251 <101390131+Poison6251@users.noreply.github.com>
Date:   Wed May 1 16:49:57 2024 +0900

    Merge pull request #2 from JHLee0891/review

    Review 기능 완성

commit e0da24a8333e588b34e62903b49811844abcfd7b
Author: unknown <killiacaridos@gmail.com>
Date:   Wed May 1 12:06:57 2024 +0900

    HTML Font 주석처리, .Prettierrc 파일 추가

commit a6db95995c4dd18ac73fee8e23509c1de4188d62
Author: unknown <killiacaridos@gmail.com>
Date:   Wed May 1 11:32:55 2024 +0900

    HTML 폰트 수정
