+++ detail.html
@@ -1,0 +1,160 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
+
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ"
+      crossorigin="anonymous"
+    />
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
+      crossorigin="anonymous"
+    ></script>
+
+    <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"
+    />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+
+    <link rel="stylesheet" href="./style/reviewEdit.css" />
+    <link rel="stylesheet" href="./style/detail.css" />
+  </head>
+
+  <body>
+    <header class="py-3 mb-3 border-bottom">
+      <div
+        class="container-fluid d-grid gap-3 align-items-center"
+        style="grid-template-columns: 1fr 2fr"
+      >
+        <span
+          class="d-flex align-items-center col-lg-4 mb-lg-0 link-body-emphasis text-decoration-none"
+          data-bs-toggle="dropdown"
+          aria-expanded="false"
+        >
+          <span class="material-symbols-outlined"> movie </span>
+        </span>
+      </div>
+    </header>
+    <div id="content">
+      <div id="detail-content"></div>
+      <div id="review-wrap">
+        <div class="review-group">
+          <div class="input-group">
+            <input
+              id="username"
+              type="text"
+              aria-label="First name"
+              placeholder="Username"
+              class="form-control"
+            />
+            <input
+              id="password"
+              type="text"
+              aria-label="Last name"
+              placeholder="Password"
+              class="form-control"
+            />
+            <button
+              id="submit"
+              style="float: right"
+              type="button"
+              class="btn btn-dark"
+            >
+              저장
+            </button>
+          </div>
+    
+          <div class="input-review-group">
+            <textarea
+              id="review"
+              class="form-control"
+              placeholder="Review"
+              aria-label="With textarea"
+            ></textarea>
+          </div>
+    
+          <div id="review-list"></div>
+        </div>
+    
+        <div class="reviewEdit" id="reviewEdit">
+          <div class="reviewWindow">
+            <div class="input-group">
+              <input
+                id="editusername"
+                type="text"
+                aria-label="First name"
+                placeholder="Username"
+                class="form-control"
+              />
+              <input
+                id="editpassword"
+                type="text"
+                aria-label="Last name"
+                placeholder="Password"
+                class="form-control"
+              />
+            </div>
+    
+            <div class="input-group">
+              <textarea
+                id="editreview"
+                class="form-control"
+                placeholder="Review"
+                aria-label="With textarea"
+              ></textarea>
+            </div>
+            <div class="input-edit-group">
+              <button id="editsubmit" type="button" class="btn btn-dark">
+                저장
+              </button>
+              <button
+                id="editclose"
+                type="button"
+                class="btn btn-outline-secondary"
+              >
+                닫기
+              </button>
+            </div>
+          </div>
+        </div>
+    
+        <div class="reviewdelete" id="reviewdelete">
+          <div class="reviewDeleteWindow">
+            <div class="input-group">
+              <p class="deletetext">
+                리뷰를 지우시려면 비밀번호를 입력하셔야 합니다.
+              </p>
+    
+              <input
+                id="deletepassword"
+                type="text"
+                aria-label="Last name"
+                placeholder="Password"
+                class="form-control"
+              />
+            </div>
+            <div class="input-edit-group">
+              <button id="delete" type="button" class="btn btn-dark">삭제</button>
+              <button
+                id="deleteCancle"
+                type="button"
+                class="btn btn-outline-secondary"
+              >
+                닫기
+              </button>
+            </div>
+          </div>
+        </div>
+    
+    <script src="./js/review.js"></script>
+    <script type="module" src="./js/detail.js"></script>
+  </body>
+</html>
