+++ index.html
@@ -1,63 +1,79 @@
 <!DOCTYPE html>
 <html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
 
     <!-- bootstrap -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
-        crossorigin="anonymous"></script>
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ"
+      crossorigin="anonymous"
+    />
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
+      crossorigin="anonymous"
+    ></script>
+
     <!-- google icon & font-->
-    <link rel="stylesheet"
-        href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Jersey+25&display=swap" rel="stylesheet">
+    <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"
+    />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
 
-    <link rel="stylesheet" href="style.css" />
-</head>
+    <link rel="stylesheet" href="./style/style.css" />
+  </head>
 
-<body>
+  <body>
     <header class="py-3 mb-3 border-bottom">
-        <div class="container-fluid d-grid gap-3 align-items-center" style="grid-template-columns: 1fr 2fr;">
-            <span class="d-flex align-items-center col-lg-4 mb-lg-0 link-body-emphasis text-decoration-none"
-                data-bs-toggle="dropdown" aria-expanded="false">
-                <span class="material-symbols-outlined">
-                    movie
-                </span>
-            </span>
+      <div
+        class="container-fluid d-grid gap-3 align-items-center"
+        style="grid-template-columns: 1fr 2fr"
+      >
+        <span
+          class="d-flex align-items-center col-lg-4 mb-lg-0 link-body-emphasis text-decoration-none"
+          data-bs-toggle="dropdown"
+          aria-expanded="false"
+        >
+          <span id="movie-logo" class="material-symbols-outlined"> movie </span>
+        </span>
 
-            <div class="d-flex align-items-center">
-                <div class="w-100 me-3" role="search">
-                    <input type="search" class="form-control jersey-25-regular" placeholder="Search..."
-                        aria-label="Search" id="search-input">
-                </div>
-                <button class="btn btn-outline-secondary jersey-25-regular" type="button"
-                    id="search-button">Search</button>
-            </div>
+        <div class="d-flex align-items-center">
+          <div class="w-100 me-3" role="search">
+            <input
+              type="search"
+              class="form-control jersey-25-regular"
+              placeholder="Search..."
+              aria-label="Search"
+              id="search-input"
+            />
+          </div>
+          <button
+            class="btn btn-outline-secondary jersey-25-regular"
+            type="button"
+            id="search-button"
+          >
+            Search
+          </button>
         </div>
+      </div>
     </header>
     <div id="content-wrap"></div>
 
-    <!-- <a href="javascript:prevPage()" id="btn_prev">Prev</a>
-    <a href="javascript:nextPage()" id="btn_next">Next</a> -->
-
     <div id="pagination-controls">
-        <span id="btn-prev" class="material-symbols-outlined">
-            navigate_before
-        </span>
-        <div id="pagination-buttons"></div>
-        <span id="btn-next" class="material-symbols-outlined">
-            navigate_next
-        </span>
+      <span id="btn-prev" class="material-symbols-outlined">
+        navigate_before
+      </span>
+      <div id="pagination-buttons"></div>
+      <span id="btn-next" class="material-symbols-outlined">
+        navigate_next
+      </span>
     </div>
-
-    <script src="app.js"></script>
-</body>
-
+    <script type="module" src="./js/main.js"></script>
+  </body>
 </html>
