+++ js/detail.js
@@ -1,0 +1,9 @@
+import { getTMDBData, setMovieDetail } from "./movie.js"
+
+let movieId = new URLSearchParams(location.search).get("id")
+const details = await getTMDBData(`https://api.themoviedb.org/3/movie/${movieId}?language=en-US`)
+setMovieDetail(details);
+
+document.getElementById("movie-logo").addEventListener("click", () => {
+    window.location.href = `index.html`;
+});
