+++ style/reviewEdit.css
@@ -1,13 +1,9 @@
-body {
-  background-color: black;
-}
-
 .review-group {
   padding: 5% 5% 5% 5%;
   margin: 20px 20px 20px 20px;
   border-radius: 15px;
-  box-shadow: 5px 5px 5px 5px rgba(255, 255, 255, 0.383);
-  border: 5px solid rgba(0, 0, 0, 0.086);
+  box-shadow: 5px 5px 5px 5px rgba(0, 0, 0, 0.234);
+  border: 1px solid rgba(0, 0, 0, 0.086);
   background-color: white;
 }
